+++ src/main/java/com/hhplus/tdd/concert/domain/model/ConcertEvent.java
@@ -1,4 +1,4 @@
-package com.hhplus.tdd.concert;
+package com.hhplus.tdd.concert.domain.model;
 
 import lombok.Builder;
 import lombok.Getter;
